move type to coordinator
